+++ src/main/java/zone/moddev/mc/ironagefurniture/proxy/ClientProxy.java
@@ -1,9 +1,8 @@
 package zone.moddev.mc.ironagefurniture.proxy;
 
 import net.minecraftforge.fml.client.registry.RenderingRegistry;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.entity.SpriteRenderer;
 import zone.moddev.mc.ironagefurniture.client.renderer.SeatRenderer;
+import zone.moddev.mc.ironagefurniture.client.renderer.ThrownLavaLampRenderer;
 import zone.moddev.mc.ironagefurniture.registers.entities;
 
 public class ClientProxy extends CommonProxy {
@@ -12,6 +11,6 @@ public class ClientProxy extends CommonProxy {
     public void onSetupClient() {
         RenderingRegistry.registerEntityRenderingHandler(entities.SEAT.get(), SeatRenderer::new);
         RenderingRegistry.registerEntityRenderingHandler(entities.THROWN_LAVA_LAMP.get(),
-                manager -> new SpriteRenderer<>(manager, Minecraft.getInstance().getItemRenderer()));
+                ThrownLavaLampRenderer::new);
     }
 }